Rename mock data file.
